python_packages/kolibri-sync-extras-plugin/requirements-test.txt
pytest
pytest-django
pytest-env
mock
